# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZunoPilot API — mobile surface",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "https://api.zunopilot.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:4000/api",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/auth/profile": {
      "put": {
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Finish onboarding",
        "description": "Called when `tenant.onboardingCompleted` is false. Creates the workspace profile.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "businessName": {
                    "type": "string"
                  },
                  "businessCategoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated session",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Session"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorised"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "The token from `POST /auth/otp/verify`. Currently valid for 24 hours."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request was not acceptable. `message` says why, in plain words.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorised": {
        "description": "No token, or an expired one.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "This conversation is assigned to someone else."
          },
          "details": {
            "description": "Field-level problems, when the failure was a validation one"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "Session": {
        "type": "object",
        "description": "Everything a client needs on launch.",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "phone": {
                "type": "string",
                "nullable": true
              },
              "email": {
                "type": "string",
                "nullable": true
              },
              "fullName": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "enum": [
                  "OWNER",
                  "MANAGER",
                  "AGENT"
                ]
              },
              "emailVerified": {
                "type": "boolean"
              },
              "country": {
                "type": "string",
                "nullable": true,
                "example": "IN"
              }
            }
          },
          "tenant": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "businessName": {
                "type": "string"
              },
              "category": {
                "type": "string",
                "nullable": true,
                "example": "RESTAURANT"
              },
              "categoryLabel": {
                "type": "string",
                "nullable": true
              },
              "catalogueNoun": {
                "type": "string",
                "example": "Menu",
                "description": "What this business calls its catalogue. Use it in the UI."
              },
              "catalogueItemNoun": {
                "type": "string",
                "example": "Item"
              },
              "maskCustomerNumbers": {
                "type": "boolean"
              },
              "aiAgentEnabled": {
                "type": "boolean"
              },
              "onboardingCompleted": {
                "type": "boolean"
              }
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "What this seat may do, e.g. `inbox:reply`. For hiding UI, never for trusting."
          },
          "modules": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "MARKETING",
                "LEADS",
                "SUPPORT",
                "AI_AGENT",
                "ECOMMERCE",
                "KEYWORD_RULES"
              ]
            },
            "description": "Which optional modules this workspace has. A missing one 404s its routes."
          },
          "activeWorkspaceId": {
            "type": "string",
            "format": "uuid",
            "description": "The workspace this session is acting in. Same value as `tenant.id`, carried separately so a switcher can mark one row without comparing shapes."
          },
          "workspaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workspace"
            },
            "description": "Every workspace this login can reach. One entry for a normal account. Absent from a session minted before this field existed — treat that as \"cannot switch\"."
          }
        }
      },
      "Workspace": {
        "type": "object",
        "description": "One workspace this login belongs to, as a switcher needs it.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "businessName": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "roleName": {
            "type": "string",
            "nullable": true,
            "description": "The workspace’s own name for the role — \"Owner\", \"Shift lead\". Not the legacy enum, which would contradict what the Team screen shows."
          },
          "isOwner": {
            "type": "boolean"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "isSuspended": {
            "type": "boolean",
            "description": "Listed even though it cannot be entered. Hiding it would make a business vanish from the switcher with no explanation."
          },
          "isCurrent": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}